Extract action methods from fritzboxmetrics.go
